Rename providers to singular words to match class names
