+++ balancer-js/src/index.ts
@@ -13,8 +13,8 @@ export * from './modules/subgraph/subgraph.module';
 export * from './modules/sor/sor.module';
 export * from './modules/pools/pools.module';
 export * from './modules/pools/types';
-export * from './modules/data-providers/pools/static.provider';
-export * from './modules/data-providers/tokens/static.provider';
+export * from './modules/data-providers/pool/static.provider';
+export * from './modules/data-providers/token/static.provider';
 export * from './balancerErrors';
 export {
     SwapInfo,
