+++ balancer-js/src/modules/data-providers/pool/provider.interface.ts
