+++ balancer-js/src/modules/data-providers/pool/static.provider.ts
