+++ balancer-js/src/modules/data-providers/token/provider.interface.ts
