+++ balancer-js/src/modules/data-providers/token/static.provider.ts
